- implemented icon generation for Android and Apple platforms
