+++ include/mi_platforms.h
@@ -1,0 +1,5 @@
+#pragma once
+
+#include "mi_utility.h"
+
+int make_icon (const Options& options);
