+++ include/mi_utility.h
@@ -1,3 +1,23 @@
+#pragma once
+
+// We use the stb image libs for reading, resizing, and writing images for packing.
+#define STB_IMAGE_RESIZE_IMPLEMENTATION
+#define STB_IMAGE_WRITE_IMPLEMENTATION
+#define STB_IMAGE_IMPLEMENTATION
+#define STB_IMAGE_RESIZE_STATIC
+#define STB_IMAGE_WRITE_STATIC
+#define STB_IMAGE_STATIC
+#include "stb/stb_image_resize.h"
+#include "stb/stb_image_write.h"
+#include "stb/stb_image.h"
+
+#include <filesystem>
+#include <iostream>
+#include <fstream>
+#include <vector>
+#include <string>
+#include <stdint.h>
+
 #define JOIN( a, b) JOIN2(a, b)
 #define JOIN2(a, b) JOIN1(a, b)
 #define JOIN1(a, b) a##b
@@ -39,6 +59,43 @@ typedef   int8_t  s8;
 typedef  int16_t s16;
 typedef  int32_t s32;
 typedef  int64_t s64;
+
+namespace e_platform
+{
+    enum platform
+    {
+        win32,
+        osx,
+        ios,
+        android,
+        COUNT
+    };
+}
+typedef e_platform::platform Platform;
+
+const std::string platform_names[e_platform::COUNT] =
+{
+    "win32",
+    "osx",
+    "ios",
+    "android"
+};
+
+struct Argument
+{
+    std::string name;
+    std::vector<std::string> params; // Optional!
+};
+
+struct Options
+{
+    std::vector<int> sizes;
+    std::vector<std::string> input;
+    std::string output;
+    std::string contents;
+    bool resize = false;
+    Platform platform = e_platform::win32;
+};
 
 #define ERROR(...)                       \
 do                                       \
