+++ source/mi_main.cpp
@@ -4,17 +4,6 @@
 #endif
 #endif
 
-// We use the stb image libs for reading, resizing, and writing images for packing.
-#define STB_IMAGE_RESIZE_IMPLEMENTATION
-#define STB_IMAGE_WRITE_IMPLEMENTATION
-#define STB_IMAGE_IMPLEMENTATION
-#define STB_IMAGE_RESIZE_STATIC
-#define STB_IMAGE_WRITE_STATIC
-#define STB_IMAGE_STATIC
-#include <stb/stb_image_resize.h>
-#include <stb/stb_image_write.h>
-#include <stb/stb_image.h>
-
 #include <filesystem>
 #include <iostream>
 #include <fstream>
@@ -34,26 +23,14 @@ constexpr const char* MAKEICON_HELP_MESSAGE =
 "    -sizes: ...  [Required]  Comma-separated input size(s) of icon image to generate for the output icon.\n"
 "    -input: ...  [Required]  Comma-separated input image(s) and/or directories and/or .txt files containing file names to be used to generate the icon sizes.\n"
 "    -resize      [Optional]  Whether to resize input images to match the specified sizes.\n"
+"    -platform    [Optional]  Platform to generate icons for. Options are win32, osx, ios, android. Defaults to win32.\n"
 // "    -detail      [Optional]  Prints out detailed information on what the program is doing.\n"
 "    -version     [Optional]  Prints out the current version number of the makeicon binary and exits.\n"
 "    -help        [Optional]  Prints out this help/usage message for the program and exits.\n"
 "     output      [Required]  The name of the icon that will be generated by the program.\n";
 
-struct Argument
-{
-    std::string name;
-    std::vector<std::string> params; // Optional!
-};
-struct Options
-{
-    std::vector<int> sizes;
-    std::vector<std::string> input;
-    std::string output;
-    bool resize = false;
-};
-
-#include "mi_utility.cpp"
-#include "mi_win32.cpp"
+#include "mi_utility.h"
+#include "mi_platforms.h"
 
 static void print_version_message ()
 {
@@ -113,8 +90,15 @@ int main (int argc, char** argv)
                 }
                 else if (arg.name == "sizes")
                 {
-                    for (auto& param: arg.params) options.sizes.push_back(std::stoi(param));
-                    if (options.sizes.empty()) ERROR("No sizes provided with -sizes argument!");
+                    for (auto& param: arg.params)
+                    {
+                        // allows sizes to be specified through a json file (used for apple icon generation)
+                        if(param.find(".json") != -1)
+                            options.contents = param;
+                        else
+                            options.sizes.push_back(std::stoi(param));
+                    }
+                    if (options.sizes.empty() && options.contents.empty()) ERROR("No sizes provided with -sizes argument!");
                 }
                 else if (arg.name == "input")
                 {
@@ -160,6 +144,18 @@ int main (int argc, char** argv)
                     }
                     if (options.input.empty()) ERROR("No input provided with -input argument!");
                 }
+                else if (arg.name == "platform")
+                {
+                    std::string platform = arg.params[0];
+                    for(size_t i = 0; i < e_platform::COUNT; ++i)
+                    {
+                        if(platform == platform_names[i])
+                        {
+                            options.platform = (Platform)i;
+                            break;
+                        }
+                    }
+                }
                 else if (arg.name == "version")
                 {
                     print_version_message();
@@ -192,7 +188,7 @@ int main (int argc, char** argv)
     }
 
     // Ensure the options structure is populated with valid data. If there are no inputs, sizes, or ouput we cannot run!
-    if (options.sizes.empty())  ERROR("No icon sizes provided! Specify sizes using: -sizes:x,y,z,w...");
+    if (options.sizes.empty() && options.contents.empty())  ERROR("No icon sizes provided! Specify sizes using: -sizes:x,y,z,w...");
     if (options.input.empty())  ERROR("No input images provided! Specify input using: -input:x,y,z,w...");
     if (options.output.empty()) ERROR("No output name provded! Specify output name like so: makeicon ... outputname.ico");
     // The maximum size allows in an ICO file is 256x256! We also check for 0 or less as that would not be valid either...
@@ -206,5 +202,5 @@ int main (int argc, char** argv)
     std::sort(options.input.begin(), options.input.end());
 
     // Takes the populated options structure and uses those options to generate an icon for the desired platform.
-    return make_icon_win32(options);
+    return make_icon(options);
 }
